move block components that aren't public outside of module to components.rs file
